Write tests for description and details components
Made progress in setting up better testing. In the future, I plan to
return to the other tests and refactor based on the format I used in
these tests.

Among other things, learned how to use wrapper.debug() to display the
wrapper as the tests are using it. Still need to understand the testing
process better.